requirements.txt
torch
transformers
diffusers
Pillow
pdf2image
pytesseract
olmocr
torch==2.7.0
torchvision
torchaudio